% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcancertemp-package.R
\docType{package}
\name{microcancertemp}
\alias{microcancertemp}
\alias{microcancertemp-package}
\title{microcancertemp: A temporary package for micro-simulation for cancers}
\description{
microcancertemp: A temporary package for micro-simulation for cancers
}
